+++ .gitignore
@@ -4,3 +4,7 @@
 
 /gen/
 /out/
+
+/build/
+
+/local.properties
