--- AndroidManifest.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.yahoo.mobile.client.android.demo.rangeseekbar"
-          android:versionCode="1"
-          android:versionName="1.0">
-    <uses-sdk android:minSdkVersion="19"/>
-    <application
-            android:label="@string/app_name"
-            android:icon="@drawable/ic_launcher">
-        <activity
-                android:name="DemoActivity"
-                android:label="@string/app_name">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
-                <category android:name="android.intent.category.LAUNCHER"/>
-            </intent-filter>
-        </activity>
-    </application>
-</manifest>
