+++ rangeseekbar-sample/src/main/res/layout/main.xml
@@ -1,0 +1,68 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              xmlns:rsb="http://schemas.android.com/apk/res-auto"
+              android:orientation="vertical"
+              android:layout_width="fill_parent"
+              android:layout_height="fill_parent"
+              android:padding="5dp"
+        >
+
+    <TextView
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:text="Range seek bar initialized in code with ranges set"
+            />
+
+    <LinearLayout android:id="@+id/seekbar_placeholder"
+                  android:orientation="vertical"
+                  android:layout_width="fill_parent"
+                  android:layout_height="wrap_content"/>
+
+    <TextView
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:text="Range seek bar from xml with default range"
+            />
+
+    <com.yahoo.mobile.client.android.util.RangeSeekBar android:layout_width="fill_parent"
+                                                       android:layout_height="wrap_content"/>
+
+    <TextView
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:text="Range seek bar from xml with ranges set"
+            />
+
+    <com.yahoo.mobile.client.android.util.RangeSeekBar android:layout_width="fill_parent"
+                                                       android:layout_height="wrap_content"
+                                                       rsb:absoluteMinValue="20"
+                                                       rsb:absoluteMaxValue="150"
+            />
+
+    <TextView
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:text="Range seek bar from xml with ranges set and double values"
+            />
+
+    <com.yahoo.mobile.client.android.util.RangeSeekBar android:layout_width="fill_parent"
+                                                       android:layout_height="wrap_content"
+                                                       rsb:absoluteMinValue="15.23"
+                                                       rsb:absoluteMaxValue="140.54"
+            />
+
+    <TextView
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:text="Range seek bar from xml with single thumb to true"
+            />
+
+    <com.yahoo.mobile.client.android.util.RangeSeekBar android:layout_width="fill_parent"
+                                                       android:layout_height="wrap_content"
+                                                       rsb:absoluteMinValue="20"
+                                                       rsb:absoluteMaxValue="150"
+                                                       rsb:singleThumb="true"
+            />
+
+</LinearLayout>
+
