+++ rangeseekbar/src/main/java/com/yahoo/mobile/client/android/util/PixelUtil.java
