+++ rangeseekbar/src/main/res/drawable-hdpi/ic_launcher.png
