+++ rangeseekbar/src/main/res/drawable-xhdpi/seek_thumb_disabled.png
