+++ rangeseekbar/src/main/res/drawable-xhdpi/seek_thumb_normal.png
