+++ rangeseekbar/src/main/res/drawable-xhdpi/seek_thumb_pressed.png
