+++ rangeseekbar/src/main/res/values/attrs.xml
