+++ rangeseekbar/src/main/res/values/strings.xml
