+++ rangeseekbar/src/test/java/com/yahoo/mobile/client/android/util/RangeSeekBarTest.java
